--- Untitled.txt
@@ -1,1 +1,0 @@
-Grant Ray
