--- src/IceCream.java
@@ -1,10 +1,0 @@
-package interfaces.food;
-
-public class IceCream {
-	setName("Chicken_Fingers");
-	setCost(0.20);
-	setStoreCost(3.00);
-	setQuantity(5);
-	setExpiration(21);
-	setDesirability(1);
-}
