+++ src/interfaces/food/Almonds.java
@@ -4,7 +4,7 @@ public class Almonds extends Food {
 	public Almonds(){
 		setName("Almonds");
 		setCost(3.00);
-		SetStoreCost(5.00);
+		setStoreCost(5.00);
 		setQuantity(10);
 		setExpiration(60);
 		setDesirability(1);
