+++ src/interfaces/food/Blueberries.java
@@ -1,16 +1,12 @@
-package org.millardps.StoreSim;
-
-import interfaces.food.Food;
-
+package interfaces.food;
 
 public class Blueberries extends Food {
-	public Blueberries(){
+	public Blueberries() {
 		setDesirability(1);
 		setExpiration(7);
 		setName("Blueberries");
 		setCost(3.99);
 		setStoreCost(1.50);
 		setQuantity(5);
-		}
+	}
 }
-//
