+++ src/interfaces/food/Carrots.java
@@ -1,13 +1,12 @@
 package interfaces.food;
 
 public class Carrots extends Food {
-	public Carrots(){
+	public Carrots() {
 		setName("Carrots");
 		setCost(0.30);
 		setStoreCost(1.50);
 		setQuantity(30);
 		setExpiration(14);
-		setDesirability(1);	
-		
+		setDesirability(1);
 	}
 }
