+++ src/interfaces/food/HotPockets.java
@@ -1,12 +1,12 @@
 package interfaces.food;
 
-public class HotPockets extends Food{
-	public void HotPockets(){
+public class HotPockets extends Food {
+	public void HotPockets() {
 		setName("Hot Pockets");
 		setCost(2);
 		setStoreCost(1);
 		setQuantity(25);
-		setExpiration(3); //days
+		setExpiration(3);
 		setDesirability(1);
-			}
+	}
 }
