+++ src/interfaces/food/IceCream.java
@@ -1,0 +1,12 @@
+package interfaces.food;
+
+public class IceCream extends Food {
+	public void IceCream() {
+		setName("Chicken_Fingers");
+		setCost(0.20);
+		setStoreCost(3.00);
+		setQuantity(5);
+		setExpiration(21);
+		setDesirability(1);
+	}
+}
