+++ src/interfaces/food/PancakeMix.java
@@ -1,6 +1,7 @@
+package interfaces.food;
 
-public class PancakeMix extends food{
-	public PancakeMix{
+public class PancakeMix extends Food{
+	public PancakeMix() {
 		setName("Pancake Mix");
 		setCost(3.99);
 		setStoreCost(1.72);
