+++ src/interfaces/food/Pie.java
@@ -1,0 +1,12 @@
+package interfaces.food;
+
+public class Pie extends Food {
+	public void Pie() {
+		setDesirability(1);
+		setCost(5.00);
+		setStoreCost(7.29);
+		setQuantity(21);
+		setExpiration(7);
+		setName("Pie");
+	}
+}
