+++ src/interfaces/food/PizzaRolls.java
@@ -1,13 +1,12 @@
 package interfaces.food;
 
 public class PizzaRolls extends Food {
-	
-	public void PizzaRolls(){
+	public void PizzaRolls() {
 		setName("Pizza Rolls");
 		setCost(5);
 		setStoreCost(2);
 		setQuantity(20);
-		setExpiration(5); // days
+		setExpiration(5);
 		setDesirability(2);
 	}
 }
