+++ src/interfaces/food/Strawberries.java
@@ -1,16 +1,12 @@
-package org.millardps.StoreSim;
-
-import interfaces.food.Food;
-
+package interfaces.food;
 
 public class Strawberries extends Food {
-	public Strawberries(){
+	public Strawberries() {
 		setDesirability(2);
 		setExpiration(7);
 		setName("Strawberries");
 		setCost(4.99);
 		setStoreCost(1.50);
 		setQuantity(15);
-		}
+	}
 }
-//
